requirements.txt
discord.py
python-dotenv
openai
notion-client
flask
